Fist commit - API to register companies
Here we create a project to register companies, where we enter with 3
information nome, CNPJ e Data de Abertura.

We have a login page that are needed so the user can register, edit,
delete and read all comapanies in the system.

We are using the MVC pattern and only Tomcat, without the help of
frameworks as Springboot, the goal is to learn what happens in thoses
frameworks under the hood.